eventgoapp/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
idna==3.10
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
